abstracting all slack methods to a separate file
